# payment-service/docs/swagger.yaml
definitions:
  dto.CreatePaymentRequest:
    properties:
      amount:
        type: number
      email:
        type: string
    required:
    - amount
    - email
    type: object
  dto.PaymentResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
  dto.UpdatePaymentRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
info:
  contact: {}
paths:
  /payments:
    get:
      description: Get all payments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PaymentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all payments
      tags:
      - Payments
    post:
      consumes:
      - application/json
      description: Create a new payment
      parameters:
      - description: Create Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new payment
      tags:
      - Payments
  /payments/{id}:
    delete:
      description: Delete payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete payment by ID
      tags:
      - Payments
    get:
      description: Get payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payment by ID
      tags:
      - Payments
    put:
      consumes:
      - application/json
      description: Update payment status by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update payment by ID
      tags:
      - Payments
swagger: "2.0"